+++ tests/operators/test_node_pod_operator.py
@@ -1,5 +1,8 @@
 import unittest
 
+from airflow.kubernetes.pod_generator import PodGenerator
+from kubernetes.client.models.v1_env_var import V1EnvVar
+
 from kedro_airflow_k8s.operators.node_pod import NodePodOperator
 
 
@@ -35,6 +38,7 @@ class TestNodePodOperator(unittest.TestCase):
             annotations={"iam.amazonaws.com/role": "airflow"},
             pipeline="data_science_pipeline",
             parameters="ds:{{ ds }}",
+            env_vars={"var1": "var1value"},
         )
 
         pod = task.create_pod_request_obj()
@@ -77,6 +81,7 @@ class TestNodePodOperator(unittest.TestCase):
 
         assert pod.spec.service_account_name == "default"
         assert len(pod.spec.image_pull_secrets) == 0
+        assert container.env[0] == V1EnvVar(name="var1", value="var1value")
 
     def test_task_create_no_limits_and_requests(self):
         task = NodePodOperator(
@@ -97,7 +102,7 @@ class TestNodePodOperator(unittest.TestCase):
         container = pod.spec.containers[0]
         assert container.resources.limits == {}
         assert container.resources.requests == {}
-        assert pod.spec.node_selector is None
+        assert not pod.spec.node_selector
 
     def test_task_with_service_account(self):
         task = NodePodOperator(
@@ -119,3 +124,31 @@ class TestNodePodOperator(unittest.TestCase):
         assert len(pod.spec.image_pull_secrets) == 2
         assert pod.spec.image_pull_secrets[0].name == "top"
         assert pod.spec.image_pull_secrets[1].name == "secret"
+
+    def test_task_with_custom_k8s_pod_template(self):
+        task = NodePodOperator(
+            node_name="test_node_name",
+            namespace="airflow",
+            pvc_name="shared_storage",
+            image="registry.gitlab.com/test_image",
+            image_pull_policy="Always",
+            env="test-pipelines",
+            task_id="test-node-name",
+            volume_owner=100,
+            mlflow_enabled=False,
+            kubernetes_pod_template=f"""
+type: Pod
+metadata:
+  name: {PodGenerator.make_unique_pod_id('test-node-name')}'
+  labels:
+    test: mylabel
+spec:
+  containers:
+    - name: base
+""",
+        )
+        pod = task.create_pod_request_obj()
+
+        assert pod.metadata.name.startswith("test-node-name")
+        assert "test-node-name" != pod.metadata.name
+        assert pod.metadata.labels["test"] == "mylabel"